added formatter
Added Prettier package as a development dependecy in the project that will auto-format the entire project.
During development, use `npm run dev`, which will run prettier on the project then run nodemon to run the project for development.
During deployment, use `npm run dep`, which will run prettier on the project then run pm2 to run a background process of the project for deployment.
To stop the nodejs server during deployment, use `npm run stop` to stop the pm2 background process of the project.